[#1634] Make -testnames find tests from Multi-level parent-child suites; Make -xmlpathinjar support <suite-files>
