+++ src/main/java/org/testng/TestNG.java
@@ -57,6 +57,8 @@ import org.testng.xml.XmlTest;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.ParameterException;
+
+import org.testng.xml.internal.TestNamesHelper;
 import org.testng.xml.internal.XmlSuiteUtils;
 
 import static org.testng.internal.Utils.defaultIfStringEmpty;
@@ -298,14 +300,14 @@ public class TestNG {
           continue;
         }
         // If test names were specified, only run these test names
-        XmlSuiteUtils xmlSuiteUitls = new XmlSuiteUtils();
-        xmlSuiteUitls.cloneIfContainsTestsWithNamesMatchingAny(s, m_testNames);
-        List<String> missMatchedTestname = xmlSuiteUitls.getMissMatchedTestNames(m_testNames);
-        if (CollectionUtils.hasElements(missMatchedTestname)) {
+        TestNamesHelper testNamesHelper = new TestNamesHelper();
+        testNamesHelper.cloneIfContainsTestsWithNamesMatchingAny(s, m_testNames);
+        List<String> missMatchedTestname = testNamesHelper.getMissMatchedTestNames(m_testNames);
+        if (!missMatchedTestname.isEmpty()) {
           throw new TestNGException("The test(s) <" + Arrays.toString(missMatchedTestname.toArray())
                   + "> cannot be found.");
         } 
-        m_suites.addAll(xmlSuiteUitls.getCloneSuite());
+        m_suites.addAll(testNamesHelper.getCloneSuite());
       }
     } catch (IOException e) {
       e.printStackTrace(System.out);
