+++ src/main/java/org/testng/xml/internal/XmlSuiteUtils.java
@@ -7,7 +7,6 @@ import org.testng.xml.XmlClass;
 import org.testng.xml.XmlSuite;
 import org.testng.xml.XmlTest;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
@@ -15,29 +14,6 @@ import java.util.Set;
  * A utility class to work with {@link XmlSuite}
  */
 public final class XmlSuiteUtils {
-    private final List<XmlSuite> cloneSuites = Lists.newArrayList();
-    private final List<String> matchedTestNames = Lists.newArrayList();
-    private final List<XmlTest> matchedTests = Lists.newArrayList();
-
-    /**
-     * Recursive search the given testNames from the current {@link XmlSuite} and its child suites.
-     *
-     * @param xmlSuite  The {@link XmlSuite} to work with.
-     * @param testNames The list of testnames to iterate through
-     */
-    public void cloneIfContainsTestsWithNamesMatchingAny(XmlSuite xmlSuite, List<String> testNames) {
-        if (testNames == null || testNames.isEmpty()) {
-            throw new TestNGException("Please provide a valid list of names to check.");
-        }
-        
-        //Start searching in the current suite.
-        addIfNotNull(cloneIfSuiteContainTestsWithNamesMatchingAny(xmlSuite, testNames));
-        
-        //Search through all the child suites.
-        for (XmlSuite suite : xmlSuite.getChildSuites()) {
-            cloneIfContainsTestsWithNamesMatchingAny(suite, testNames);
-        }
-    }
 
     /**
      * A validator that runs through the list of suites and checks if each of the suites contains
@@ -70,35 +46,6 @@ public final class XmlSuiteUtils {
         xmlTest.setXmlClasses(constructXmlClassesUsing(classes));
         return xmlSuite;
     }
-    
-    public List<XmlSuite> getCloneSuite() {
-        return cloneSuites;
-    }
-
-    /**
-     * @param testNames input from m_testNames
-     * 
-     */
-    public List<String> getMissMatchedTestNames(List<String> testNames){
-        List<String> tmpTestNames = Lists.newArrayList();
-        tmpTestNames.addAll(testNames);
-        Iterator<String> testNameIterator = tmpTestNames.iterator();
-        while (testNameIterator.hasNext()) {
-            String testName = testNameIterator.next();
-            if (matchedTestNames.contains(testName)) {
-                testNameIterator.remove();
-            }
-        }
-        return tmpTestNames;        
-    }
-
-    public List<XmlTest> getMatchedTests() {
-        return matchedTests;
-    }
-
-    public List<String> getMatchedTestNames() {
-        return matchedTestNames;
-    }
 
     /**
      * Ensures that the current suite doesn't contain any duplicate {@link XmlTest} instances.
@@ -116,12 +63,6 @@ public final class XmlSuiteUtils {
         }
     }
     
-    private void addIfNotNull(XmlSuite xmlSuite) {
-        if (xmlSuite != null) {
-            cloneSuites.add(xmlSuite);
-        }
-    }
-
     private static List<XmlClass> constructXmlClassesUsing(List<String> classes) {
         List<XmlClass> xmlClasses = Lists.newLinkedList();
         for (String cls : classes) {
@@ -129,28 +70,6 @@ public final class XmlSuiteUtils {
             xmlClasses.add(xmlClass);
         }
         return xmlClasses;
-    }
-
-    private XmlSuite cloneIfSuiteContainTestsWithNamesMatchingAny(XmlSuite suite, List<String> testNames) {
-        List<XmlTest> tests = Lists.newLinkedList();
-        for (XmlTest xt : suite.getTests()) {
-            if (xt.nameMatchesAny(testNames)) {
-                tests.add(xt);
-                matchedTestNames.add(xt.getName());
-                matchedTests.add(xt);
-            }
-        }
-        if (tests.isEmpty()) {
-            return null;
-        }
-        return cleanClone(suite, tests);
-    }
-
-    private static XmlSuite cleanClone(XmlSuite xmlSuite, List<XmlTest> tests) {
-        XmlSuite result = (XmlSuite) xmlSuite.clone();
-        result.getTests().clear();
-        result.getTests().addAll(tests);
-        return result;
     }
 
     private static void adjustSuiteNamesToEnsureUniqueness(List<XmlSuite> suites, Set<String> names) {
